SQL scripts!

+++ scripts/create_tables/create_phf_game_goalies.sql
@@ -7,11 +7,11 @@ create table phf_game_goalies (
     jersey int, 
     game_id int, 
     time_on_ice varchar(10), 
-    minutes_played float, 
+    minutes_played DECIMAL(5, 1), 
     shots_against int, 
     saves int, 
     goals_allowed int, 
-    save_pct float, 
+    save_pct DECIMAL(4, 3), 
     penalty_minutes int, 
     goals int, 
     assists int
